requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
psycopg2-binary
beautifulsoup4


